refactor: consolidate geo utilities
