+++ src/lib/__tests__/geo.test.ts
@@ -1,8 +1,14 @@
 import { describe, it, expect } from 'vitest';
-import { calculatePolygonArea, calculatePolygonCenter, processGeoJSON } from '../geojsonUtils';
-import { mockFeatureCollection } from '../../utils/geo';
+import {
+  calculatePolygonArea,
+  calculatePolygonCenter,
+  processGeoJSON,
+  parseAndValidateGeoJSON,
+  mockFeatureCollection,
+  getLoteStyle
+} from '../geo';
 
-describe('geojsonUtils', () => {
+describe('geo utilities', () => {
   const fc = mockFeatureCollection();
   const polygon = fc.features[0].geometry.coordinates;
 
@@ -47,4 +53,33 @@ describe('geojsonUtils', () => {
     });
     expect(() => processGeoJSON(invalid)).toThrow('Arquivo GeoJSON inválido');
   });
+
+  it('parseAndValidateGeoJSON parses valid feature collection', async () => {
+    const { fc: parsed, featuresCount, bbox } = await parseAndValidateGeoJSON(JSON.stringify(fc));
+    expect(featuresCount).toBe(fc.features.length);
+    expect(parsed).toEqual(fc);
+    expect(bbox).toEqual([-46.639, -23.5485, -46.637, -23.547]);
+  });
+
+  it('parseAndValidateGeoJSON throws on invalid JSON', async () => {
+    await expect(parseAndValidateGeoJSON('invalid')).rejects.toThrow('JSON inválido');
+  });
+
+  it('parseAndValidateGeoJSON throws on invalid geometry', async () => {
+    const invalid = JSON.stringify({
+      type: 'FeatureCollection',
+      features: [
+        { type: 'Feature', geometry: { type: 'Point', coordinates: [0, 0] }, properties: {} }
+      ]
+    });
+    await expect(parseAndValidateGeoJSON(invalid)).rejects.toThrow('Somente Polygon ou MultiPolygon são aceitos');
+  });
+
+  it('getLoteStyle returns style for each status', () => {
+    expect(getLoteStyle('disponivel').color).toBe('#22c55e');
+    expect(getLoteStyle('reservado').color).toBe('#eab308');
+    expect(getLoteStyle('vendido').color).toBe('#ef4444');
+    expect(getLoteStyle('outro').color).toBe('#3b82f6');
+  });
 });
+
